Use .hpp for Objective-C++ headers
