+++ src/SFML/Window/OSX/NSImage+raw.hpp
