+++ src/SFML/Window/OSX/NSImage+raw.mm
@@ -26,7 +26,7 @@
 ////////////////////////////////////////////////////////////
 // Headers
 ////////////////////////////////////////////////////////////
-#import <SFML/Window/OSX/NSImage+raw.h>
+#import <SFML/Window/OSX/NSImage+raw.hpp>
 
 @implementation NSImage (raw)
