+++ src/SFML/Window/OSX/WindowImplDelegateProtocol.hpp
