+++ src/SFML/Window/OSX/cpp_objc_conversion.mm
@@ -26,7 +26,7 @@
 ////////////////////////////////////////////////////////////
 // Headers
 ////////////////////////////////////////////////////////////
-#import <SFML/Window/OSX/cpp_objc_conversion.h>
+#import <SFML/Window/OSX/cpp_objc_conversion.hpp>
 
 #include <SFML/System/Utf.hpp>
